update :: 순환참조 에러 해결

+++ src/main/kotlin/com/example/coblebackend/domain/user/service/DeleteUserService.kt
@@ -25,10 +25,10 @@ class DeleteUserService(
     fun execute(request: DeleteUserRequest) {
         val user = userFacade.getCurrentUser()
 
-        userFacade.checkPasswordMatch(
-            password = user.password,
-            requestPassword = request.password,
-        )
+        val checkPasswordMatch = passwordEncoder.matches(user.password, request.password)
+
+        if (!checkPasswordMatch)
+            throw PasswordMismatchedException
 
         projectRepository.deleteAllByUserId(user.id)
         likeRepository.deleteAllByUserId(user.id)
